% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroughtData-package.R
\docType{package}
\name{DroughtData-package}
\alias{DroughtData}
\alias{DroughtData-package}
\title{DroughtData: Drought Synthesis Data}
\description{
Contains data sets used in the Drought Synthesis project. This data package also contains useful metadata describing the data sets, the processing scripts used to derive the data, and a couple helper functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mountaindboz/DroughtData}
  \item Report bugs at \url{https://github.com/mountaindboz/DroughtData/issues}
}

}
\author{
\strong{Maintainer}: Dave Bosworth \email{David.Bosworth@water.ca.gov} (\href{https://orcid.org/0000-0003-0740-3390}{ORCID})

}
\keyword{internal}
