% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{drt_add_yr_assign}
\alias{drt_add_yr_assign}
\title{Add year assignment information}
\usage{
drt_add_yr_assign(df)
}
\arguments{
\item{df}{A data frame or tibble containing the data to add the year
assignment information to. Must contain a variable for \code{YearAdj}, which is
the adjusted calendar year (December-November, with December of the
previous calendar year included with the following year).}
}
\value{
A tibble containing the original data with the addition of the year
assignment information
}
\description{
Adds year assignment information to a data frame or tibble. The
information added includes the following variables for each year in the
data frame. This information is used in the analyses for the Drought
Synthesis.
\itemize{
\item \code{SVIndex} - Sacramento Valley Water Year Index
\item \code{YearType} - Water year type based on Sacramento Valley Water Year Index
(Critical, Dry, Below Normal, Above Normal, or Wet)
\item \code{Drought} - Drought/Wet/Neutral classification. Multiple Dry, Critical, and
Below Normal years in a row are a drought (D), multiple Wet or Above Normal
years in a row are a wet period (W), and years that are not on a streak are
neutral (N).
}
}
\examples{
# Add year assignment information to raw water quality data
df_wq <- drt_add_yr_assign(raw_wq_1975_2021)
dplyr::glimpse(df_wq)

}
