% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{drt_avg_data}
\alias{drt_avg_data}
\title{Calculate seasonal and/or regional averages}
\usage{
drt_avg_data(
  df,
  data_var,
  avg_type = c("season", "region", "both"),
  month.na = c("strict", "relaxed"),
  .quote = FALSE
)
}
\arguments{
\item{df}{A data frame or tibble containing the raw data to average. Must
contain variables for \code{Month}, \code{Season}, \code{Region}, and \code{YearAdj}.}

\item{data_var}{The variable name in \code{df} containing the values to be
averaged. Supports tidy or non-standard evaluation as default. Place the
variable name in quotes if the \code{.quote} argument is \code{TRUE}.}

\item{avg_type}{Use \code{"season"} for seasonal averages, \code{"region"} for regional
averages, or \code{"both"} for seasonal-regional averages}

\item{month.na}{Either \code{"strict"} to make sure all months are represented in
each season, or \code{"relaxed"} to allow for one or more missing months in a
season}

\item{.quote}{An optional argument that determines whether the function uses
tidy or non-standard evaluation. The default (\code{TRUE}) requires the variable
name in \code{data_var} to be unquoted. If this argument is set to \code{FALSE},
\code{data_var} must be quoted. Allowing for quoted variable names can be useful
to run more than one variable through the function using a character vector
of variable names.}
}
\value{
A tibble containing the averaged data
}
\description{
Calculates the seasonal, regional, or seasonal-regional averages
of raw data used for the Drought Synthesis
}
