% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{drt_color_pal}
\alias{drt_color_pal}
\alias{drt_color_pal_drought}
\alias{drt_color_pal_yrtype}
\title{Color palettes for Drought Synthesis report}
\usage{
drt_color_pal_drought(aes_type = c("fill", "color"), scale_title = "Drought")

drt_color_pal_yrtype(aes_type = c("fill", "color"), scale_title = "Year Type")
}
\arguments{
\item{aes_type}{A string that identifies the aesthetic type to apply the
palette to. Must be either \code{"fill"} or \code{"color"}.}

\item{scale_title}{(optional) A string that provides a title for the scale
of the color palette. Defaults are:
\itemize{
\item "Drought" for \code{drt_color_pal_drought()}
\item "Year Type" for \code{drt_color_pal_yrtype()}
}}
}
\value{
A discrete color scale for a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Apply custom color palettes to the figures in the Drought
Synthesis report. The palettes are based on the
\href{https://cran.r-project.org/web/packages/viridis/vignettes/intro-to-viridis.html}{viridis}
palette. Functions can only be used with \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} objects.
There are two palette options:
\itemize{
\item \code{drt_color_pal_drought()} is intended for the three Drought categories (D,
N, W)
\item \code{drt_color_pal_yrtype()} is intended for the five Year Type categories
(Critical, Dry, Below Normal, Above Normal, Wet)
}
}
\examples{
# A simple boxplot of seasonal average outflow for the Drought categories -
  # using default ggplot2 colors for the fill aesthetic
lt_outflow <- dplyr::filter(lt_avg_hydro, !is.na(Outflow))

library(ggplot2)
p <- ggplot(lt_outflow, aes(x = Season, y = Outflow, fill = Drought)) +
  geom_boxplot()
p

# Apply the custom color palette for the Drought categories
p + drt_color_pal_drought(aes_type = "fill")

# Change the default scale title
p + drt_color_pal_drought(aes_type = "fill", scale_title = "Something Else")

# Use the custom color palette for the color aesthetic instead
ggplot(lt_outflow, aes(x = Season, y = Outflow, color = Drought)) +
  geom_boxplot() +
  drt_color_pal_drought(aes_type = "color")

}
\seealso{
\code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}, \code{\link[ggplot2:scale_manual]{ggplot2::scale_color_manual()}}
}
