% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{drt_replace_rl}
\alias{drt_replace_rl}
\title{Replace values below the reporting limit}
\usage{
drt_replace_rl(df, data_var, sign_var, min_val = 1e-05, seed = 1)
}
\arguments{
\item{df}{A data frame or tibble containing the data to run the replacement
procedure on.}

\item{data_var}{The variable name in \code{df} containing the values to run the
replacement procedure on. Must be numeric with the records measured below
the reporting limit substituted with the analytical reporting limit. Can
contain a mixture of values, some measured below the reporting limit and
others that aren't.}

\item{sign_var}{The variable name in \code{df} containing the qualifier codes for
the values. The codes used to indicate whether a value was measured below
the reporting limit must start with "<".}

\item{min_val}{The minimum value allowed for the random number generation for
the values below the reporting limit. Default is 0.00001, but the user can
specify their preferred value.}

\item{seed}{The seed specification for the random number generation. Default
is 1, but the user can specify their preferred value. Allows for the random
number generation procedure to be reproducible. See the documentation for
\code{\link[base:Random]{base::set.seed()}} for more information.}
}
\value{
A tibble containing the original data with random numbers of uniform
distribution substituted for the values measured below the reporting limit.
All other values remain the same.
}
\description{
Replaces values measured below the analytical reporting limit
with a random number of uniform distribution between 0.00001 (default) and
the reporting limit. This function only replaces the values below the
reporting limit. It does not change any other values in the data.
}
\examples{
# Replace Dissolved Ammonia values below the reporting limit with a random
  # number
drt_replace_rl(raw_nutr_1975_2021, DissAmmonia, DissAmmonia_Sign)

# Use the min_val argument to set a different minimum value for the range of
  # random values generated
drt_replace_rl(raw_nutr_1975_2021, DissAmmonia, DissAmmonia_Sign, min_val = 0.0000001)

# Use the seed argument to set a different random seed
drt_replace_rl(raw_nutr_1975_2021, DissAmmonia, DissAmmonia_Sign, seed = 500)

}
\seealso{
\code{\link[stats:Uniform]{stats::runif()}}
}
