% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lt_avg_hydro}
\alias{lt_avg_hydro}
\title{Long-term seasonal averages of hydrology parameters}
\format{
data frame with 188 rows and 8 columns
\describe{
\item{YearAdj}{Adjusted calendar year - December-November, with December of
the previous calendar year included with the following year. For example,
December of 2020 is assigned year "2021" in this data set.}
\item{Season}{Winter (Dec-Feb), Spring (Mar-May), Summer (Jun-Aug), or Fall
(Sept-Nov)}
\item{SVIndex}{Sacramento Valley Water Year Index}
\item{YearType}{Water year type based on Sacramento Valley Water Year Index
(Critical, Dry, Below Normal, Above Normal, or Wet). The Fall season
(Sept-Nov) straddles two water years, but we assigned it to the previous
water year.}
\item{Drought}{Drought/Wet/Neutral classification. Multiple Dry, Critical,
and Below Normal years in a row are a drought (D), multiple Wet or Above
Normal years in a row are a wet period (W), and years that are not on a
streak are neutral (N).}
\item{Outflow}{Seasonal averages of net Delta outflow in cubic feet per
second (cfs) at Chipps Island. Data from WY 1975-2021 is from Dayflow. The
raw values for October through November 2021 were estimated by data from
the USBR Delta Outflow Computation reports.}
\item{Export}{Seasonal averages of the total amount of water diverted from
the Delta by the Federal and State governments to meet water agreements and
contracts, expressed in cubic feet per second (cfs). Data from WY 1975-2021
is from Dayflow. The raw values for October through November 2021 were
estimated by data from the USBR Delta Outflow Computation reports.}
\item{X2}{Seasonal averages of the distance from the Golden Gate Bridge to
the point where daily average salinity is 2 parts per thousand at a depth
of 1 meter off the bottom, expressed in kilometers. From Hutton et al.
(2017) for WY 1975-1996 and from Dayflow for WY 1997-2021. Raw values were
estimated using the Autoregressive Lag Model from Dayflow for October
through November 2021.}
}
}
\source{
\itemize{
\item Sacramento Valley Water Year Index:
\url{https://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST}
\item Dayflow: \url{https://data.cnra.ca.gov/dataset/dayflow}
\item USBR Delta Outflow Computation reports:
\url{https://www.usbr.gov/mp/cvo/pmdoc.html}
\item Hutton, P. H., J. S. Rath, and S. B Roy. 2017. Freshwater flow to the San
Francisco Bay-Delta estuary over nine decades (Part 1): Trend evaluation.
Hydrological Processes \strong{31}(14): 2500-2515. \url{https://doi.org/10.1002/hyp.11201}
}
}
\usage{
lt_avg_hydro
}
\description{
A long-term (1975-2021) data set containing seasonal averages of
three hydrology parameters (Delta Outflow, Delta Exports, X2). Used in the
analyses for the Drought Synthesis.
}
\keyword{datasets}
