% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lt_avg_wq}
\alias{lt_avg_wq}
\title{Long-term seasonal-regional averages of water quality parameters}
\format{
data frame with 940 rows and 9 columns
\describe{
\item{YearAdj}{Adjusted calendar year - December-November, with December of
the previous calendar year included with the following year. For example,
December of 2020 is assigned year "2021" in this data set.}
\item{Season}{Winter (Dec-Feb), Spring (Mar-May), Summer (Jun-Aug), or Fall
(Sept-Nov)}
\item{Region}{Region designation within the larger Delta and Suisun
Marsh/Bay (Confluence, North, SouthCentral, Suisun Bay, Suisun Marsh).}
\item{SVIndex}{Sacramento Valley Water Year Index}
\item{YearType}{Water year type based on Sacramento Valley Water Year Index
(Critical, Dry, Below Normal, Above Normal, or Wet). The Fall season
(Sept-Nov) straddles two water years, but we assigned it to the previous
water year.}
\item{Drought}{Drought/Wet/Neutral classification. Multiple Dry, Critical,
and Below Normal years in a row are a drought (D), multiple Wet or Above
Normal years in a row are a wet period (W), and years that are not on a
streak are neutral (N).}
\item{Temperature}{Seasonal-regional averages of water temperature in
degrees Celsius}
\item{Salinity}{Seasonal-regional averages of salinity value on the
Practical Salinity Scale}
\item{Secchi}{Seasonal-regional averages of secchi depth in centimeters}
}
}
\source{
\itemize{
\item Sacramento Valley Water Year Index:
\url{https://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST}
\item \code{discretewq} data package (version 2.3.2.9000):
\url{https://github.com/sbashevkin/discretewq/tree/dbb2638e800e637f3f2c70d47c9faad68f95db07}
}
}
\usage{
lt_avg_wq
}
\description{
A long-term (1975-2021) data set containing seasonal-regional
averages of three water quality parameters (water temperature, salinity,
and secchi depth) within the upper San Francisco Estuary (Delta). Used in
the analyses for the Drought Synthesis.
}
\keyword{datasets}
