% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lt_regional}
\alias{lt_regional}
\title{Long-term regional averages}
\format{
data frame with 235 rows and 8 columns
\describe{
\item{YearAdj}{Adjusted calendar year - December-November, with December of
the previous calendar year included with the following year. For example,
December of 2019 is the first month of 2020 in this data set.}
\item{Region}{Region designation within the larger Delta and Suisun
Marsh/Bay (Confluence, North, SouthCentral, Suisun Bay, Suisun Marsh).}
\item{SVIndex}{Sacramento Valley Water Year Index}
\item{YearType}{Water year type based on Sacramento Valley Water Year Index
(Critical, Dry, Below Normal, Above Normal, or Wet). The Fall season
(Sept-Nov) straddles two water years, but we assigned it to the previous
water year.}
\item{Drought}{Drought/Wet/Neutral classification. Multiple Dry, Critical,
and Below Normal years in a row are a drought (D), multiple Wet or Above
Normal years in a row are a wet period (W), and years that are not on a
streak are neutral (N).}
\item{Temperature}{Water temperature in degrees Celsius from the
\code{discretewq} data package. Data from EMP and FMWT collected in 2021 were
provided directly from staff who collected the data and should be
considered provisional.}
\item{Salinity}{Salinity value on the Practical Salinity Scale from the
\code{discretewq} data package. Data from EMP and FMWT collected in 2021 were
provided directly from staff who collected the data and should be
considered provisional.}
\item{Secchi}{Secchi depth in centimeters from the \code{discretewq} data
package. Data from EMP and FMWT collected in 2021 were provided directly
from staff who collected the data and should be considered provisional.}
}
}
\source{
\itemize{
\item Sacramento Valley Water Year Index: \url{https://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST}
\item \code{discretewq} data package (version 2.1.0): \url{https://github.com/sbashevkin/discretewq/tree/1af04438c961c6d51f46cc777cbb6a142c1e3b30}
}
}
\usage{
lt_regional
}
\description{
A long-term (1975-2021) integrated data set containing regional
averages of various metrics for each year. Used in the analyses for the
Drought Synthesis.
}
\keyword{datasets}
