% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_hydro_1975_2021}
\alias{raw_hydro_1975_2021}
\title{Raw daily values for hydrology and LSZ metrics for 1975-2021}
\format{
data frame with 17,137 rows and 5 columns
\describe{
\item{YearAdj}{Adjusted calendar year - December-November, with December of
the previous calendar year included with the following year. For example,
December of 1974 is assigned year "1975" in this data set.}
\item{Date}{Calendar date}
\item{Outflow}{Net Delta outflow in cubic feet per second (cfs) at Chipps
Island from Dayflow. The values for October 2020 through October 2021 are
estimated by data from the DTO station on CDEC.}
\item{Export}{Total amount of water diverted from the Delta by the Federal
and State governments to meet water agreements and contracts, expressed in
cubic feet per second (cfs). From Dayflow. No data available for October
2020 through October 2021 at this point.}
\item{X2}{The distance from the Golden Gate Bridge to the point where daily
average salinity is 2 parts per thousand at a depth of 1 meter off the
bottom, expressed in kilometers. From Hutton et. al. for WY 1975-1996, from
Dayflow for WY 1997-2020, and estimated using the Autoregressive Lag Model
from Dayflow and Outflow from the DTO station on CDEC for October 2020
through August 2021.}
}
}
\source{
\itemize{
\item Dayflow: \url{https://data.cnra.ca.gov/dataset/dayflow}
\item DTO data from CDEC: \url{https://cdec.water.ca.gov/dynamicapp/staMeta?station_id=DTO}
}
}
\usage{
raw_hydro_1975_2021
}
\description{
A data set containing raw daily values of the hydrology and Low
Salinity Zone (LSZ) metrics for 1975-2021. Used in the analyses for the
Drought Synthesis.
}
\keyword{datasets}
