% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_hydro_1975_2022}
\alias{raw_hydro_1975_2022}
\title{Raw daily values of hydrology and LSZ metrics for 1975-2022}
\format{
data frame with 17,532 rows and 12 columns
\describe{
\item{YearAdj}{Adjusted calendar year - December-November, with December of
the previous calendar year included with the following year. For example,
December of 2020 is assigned year "2021" in this data set.}
\item{Season}{Winter (Dec-Feb), Spring (Mar-May), Summer (Jun-Aug), or Fall
(Sept-Nov)}
\item{Date}{Calendar date of the value in yyyy-mm-dd}
\item{InflowSacR}{Sacramento River inflow to the Delta in cubic feet per
second (cfs). Data from WY 1975-2022 is from Dayflow. The values for
October through November 2022 are from the Sacramento River at Freeport
station (11447650) accessed through the USGS National Water Information
System.}
\item{InflowYolo}{Yolo Bypass inflow to the Delta in cubic feet per
second (cfs). Data from WY 1975-2022 is from Dayflow.}
\item{InflowEast}{Eastern Delta inflow in cubic feet per second (cfs).
Includes inflows from the San Joaquin, Cosumnes and Mokelumne Rivers, and
miscellaneous streams from the northeast, east, and southeast. Data from WY
1975-2022 is from Dayflow.}
\item{InflowTotal}{Total Delta inflow in cubic feet per second (cfs). Data
from WY 1975-2022 is from Dayflow.}
\item{Outflow}{Net Delta outflow in cubic feet per second (cfs) at Chipps
Island. Data from WY 1975-2022 is from Dayflow. The values for October
through November 2022 are estimated by data from the USBR Delta Outflow
Computation reports.}
\item{Export}{Total amount of water diverted from the Delta by the Federal
and State governments to meet water agreements and contracts, expressed in
cubic feet per second (cfs). Data from WY 1975-2022 is from Dayflow. The
values for October through November 2022 are estimated by data from the
USBR Delta Outflow Computation reports.}
\item{X2}{The distance from the Golden Gate Bridge to the point where daily
average salinity is 2 parts per thousand at a depth of 1 meter off the
bottom, expressed in kilometers. From Hutton et al. (2017) for WY
1975-1996, from Dayflow for WY 1997-2022, and estimated using the
Autoregressive Lag Model from Dayflow for October through November 2022.}
\item{TotalUSGSOutflow}{Daily averages of combined tidally-filtered flow
data in cubic feet per second (cfs) from four USGS stations: Sacramento
River at Rio Vista CA (11455420), San Joaquin River at Jersey Point CA
(11337190), Threemile Slough near Rio Vista CA (11337080), and Dutch Slough
below Jersey Island Road at Jersey Island (11313433). Regarded to be an
estimate of daily net Delta outflow. Data was obtained from the USGS
National Water Information System. Some of the data is considered
provisional.}
\item{CacheFlow}{Daily averages of tidally-filtered flow data for Cache
Slough in cubic feet per second (cfs). Data is from two USGS stations:
Cache Slough at Ryer Island (11455350) from 3/6/2003 to 4/26/2019, and
Cache Slough above Ryer Island Ferry near Rio Vista CA (11455385) from
4/27/2019 to 11/30/2021. Data was obtained from the USGS National Water
Information System. Some of the data is considered provisional.}
}
}
\source{
\itemize{
\item Dayflow: \url{https://data.cnra.ca.gov/dataset/dayflow}
\item USBR Delta Outflow Computation reports:
\url{https://www.usbr.gov/mp/cvo/pmdoc.html}
\item Hutton, P. H., J. S. Rath, and S. B Roy. 2017. Freshwater flow to the San
Francisco Bay-Delta estuary over nine decades (Part 1): Trend evaluation.
Hydrological Processes \strong{31}(14): 2500-2515. \url{https://doi.org/10.1002/hyp.11201}.
\item USGS National Water Information System (NWIS):
\url{https://nwis.waterdata.usgs.gov/nwis}
}
}
\usage{
raw_hydro_1975_2022
}
\description{
A data set containing raw daily values of the hydrology and Low
Salinity Zone (LSZ) metrics for 1975-2022 in the upper San Francisco
Estuary (Delta). Seasonal averages of the \code{Outflow}, \code{Export}, and \code{X2}
data are in the \code{\link{lt_avg_hydro}} data set. Used in the analyses for the
Drought Synthesis.
}
\keyword{datasets}
