% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_nutr_1975_2021}
\alias{raw_nutr_1975_2021}
\title{Raw values of nutrient concentrations for 1975-2021}
\format{
data frame with 13,397 rows and 17 columns
\describe{
\item{Source}{Name of the source dataset. Either EMP (DWR's Environmental
Monitoring Program), USGS_SFBS (USGS's San Francisco Bay Water Quality
Survey), or USGS_CAWSC (USGS's California Water Science Center)}
\item{Station}{Location where measurement was collected}
\item{Latitude}{Latitude of \code{Station} in Decimal Degrees (WGS 84 Datum)}
\item{Longitude}{Longitude of \code{Station} in Decimal Degrees (WGS 84 Datum)}
\item{Region}{Region designation of \code{Station} within the larger Delta and
Suisun Marsh/Bay (Confluence, North, SouthCentral, Suisun Bay, Suisun
Marsh). The Drought Synthesis team categorized \code{SubRegion}s into broader
\code{Region}s using \href{https://github.com/mountaindboz/DroughtData/blob/master/data-raw/Rosies_regions.csv}{this crosswalk}
.}
\item{SubRegion}{Subregion designation of \code{Station} as defined by the
\code{R_EDSM_Subregions_Mahardja_FLOAT} shapefile from the
\href{https://github.com/InteragencyEcologicalProgram/deltamapr}{\code{deltamapr}}
package.}
\item{YearAdj}{Adjusted calendar year - December-November, with December of
the previous calendar year included with the following year. For example,
December of 2020 is assigned year "2021" in this data set.}
\item{Season}{Winter (Dec-Feb), Spring (Mar-May), Summer (Jun-Aug), or Fall
(Sept-Nov)}
\item{Month}{Integer representing the month}
\item{Date}{Calendar date of the measurement}
\item{Datetime}{Date and time of the measurement}
\item{DissAmmonia_Sign}{A symbol representing whether the Dissolved Ammonia
value is below the reporting limit or above it. Contains one of four
symbols: "<", "< (estimated)", "=", or "= (unreliable)". See the section
below for more information about what these symbols represent.}
\item{DissAmmonia}{Dissolved Ammonia value in mg/L as N}
\item{DissNitrateNitrite_Sign}{A symbol representing whether the Dissolved
Nitrate + Nitrite value is below the reporting limit or above it. Contains
one of four symbols: "<", "< (estimated)", "=", or "= (unreliable)". See
the section below for more information about what these symbols represent.}
\item{DissNitrateNitrite}{Dissolved Nitrate + Nitrite value in mg/L as N}
\item{DissOrthophos_Sign}{A symbol representing whether the Dissolved
Ortho-phosphate value is below the reporting limit or above it. Contains
one of four symbols: "<", "< (estimated)", "=", or "= (unreliable)". See
the section below for more information about what these symbols represent.}
\item{DissOrthophos}{Dissolved Ortho-phosphate value in mg/L as P}
}
}
\source{
\itemize{
\item \code{discretewq} data package (version 2.3.1):
\url{https://github.com/sbashevkin/discretewq/tree/v2.3.1}
\item Water Quality Portal hosted by the National Water Quality Monitoring
Council: \url{https://www.waterqualitydata.us/}
}
}
\usage{
raw_nutr_1975_2021
}
\description{
A long-term data set containing raw values of three nutrient
parameters (Dissolved Ammonia, Dissolved Nitrate + Nitrite, and Dissolved
Ortho-phosphate) for 1975-2021. Used in the analyses for the Drought
Synthesis.

Most of the nutrient data from EMP (DWR's Environmental Monitoring Program)
and all of the nutrient data from USGS_SFBS (USGS's San Francisco Bay Water
Quality Survey) are from the \code{discretewq} data package. Data from EMP
collected in 2021 was provided directly from staff who collected the data
and should be considered provisional. Data from USGS_CAWSC were downloaded
directly from the Water Quality Portal using the \code{dataRetrieval} R package.
Some of the data from USGS_CAWSC is also considered provisional.

This data set contains only one data point per \code{Station} and \code{Date} and
excludes any data points that have modified z-scores greater than 15
grouped by \code{SubRegion}. Additionally, samples collected in the Suisun Marsh
\code{Region} were excluded from this data set due to a large gap in the
long-term record.
}
\section{More information about the \verb{_Sign} variables}{

For the variables that have the _Sign suffix, the symbols in these variables
represent four conditions of the value contained in its corresponding result
variable.
\itemize{
\item "<" - The value is below the Reporting Limit (RL) with the value in the
corresponding result variable equal to the RL.
\item "< (estimated)" - The value is below the RL, but the RL is unknown. The
value in the corresponding result variable is the estimated RL value.
\item "=" - The value is above the RL with the value in the corresponding result
variable equal to the actual value measured by the laboratory. An \code{NA} value
in the corresponding result variable indicates that the value is missing or
wasn't collected.
\item "= (unreliable)" - The value in the corresponding result variable was
determined to be an outlier and is excluded from the data set.
}
}

\keyword{datasets}
