% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{st_hydro_2021}
\alias{st_hydro_2021}
\title{Daily hydrology and LSZ values for water year 2021}
\format{
data frame with 396 rows and 3 columns
\describe{
\item{Date}{Calendar date}
\item{Outflow}{Estimated net Delta outflow in cubic feet per second (cfs)
from the DTO station on CDEC.}
\item{X2}{The distance from the Golden Gate Bridge to the point where daily
average salinity is 2 parts per thousand at a depth of 1 meter off the
bottom, expressed in kilometers. Estimated using the Autoregressive Lag
Model from Dayflow and Outflow from the DTO station on CDEC.}
}
}
\source{
\itemize{
\item Dayflow: \url{https://data.cnra.ca.gov/dataset/dayflow}
\item DTO data from CDEC: \url{https://cdec.water.ca.gov/dynamicapp/staMeta?station_id=DTO}
}
}
\usage{
st_hydro_2021
}
\description{
Daily values for the hydrology and Low Salinity Zone (LSZ)
metrics for 2021 (Oct 2020 - Oct 2021). Used in the analyses for the
Drought Synthesis. \strong{NOTE:} \code{Export} does not exist in this data set since
DAYFLOW data isn't published for WY 2021 at this point.
}
\keyword{datasets}
