% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lt_seasonal}
\alias{lt_seasonal}
\title{Long-term seasonal averages}
\format{
data frame with 188 rows and 11 columns
\describe{
\item{Year}{Adjusted calendar year - December-November, with December of
the previous calendar year included with the following year. For example,
December of 2019 is the first month of 2020 in this data set.}
\item{Season}{Winter (Dec-Feb), Spring (Mar-May), Summer (Jun-Aug), or Fall
(Sept-Nov)}
\item{SVIndex}{Sacramento Valley Water Year Index}
\item{YearType}{Water year type based on Sacramento Valley Water Year Index
(Critical, Dry, Below Normal, Above Normal, or Wet). The Fall season
(Sept-Nov) straddles two water years, but we assigned it to the previous
water year.}
\item{Drought}{Drought/Wet/Neutral classification. Multiple Dry, Critical,
and Below Normal years in a row are a drought (D), multiple Wet or Above
Normal years in a row are a wet period (W), and years that are not on a
streak are neutral (N).}
\item{Outflow}{Net Delta outflow in cubic feet per second (cfs) at Chipps
Island from Dayflow. The values for water year 2021 are estimated by data
from the DTO station on CDEC.}
\item{X2}{The distance from the Golden Gate Bridge to the point where daily
average salinity is 2 parts per thousand at a depth of 1 meter off the
bottom, expressed in kilometers. From Dayflow for recent years (WY
1997-2020), Hutton et. al. for earlier years.}
\item{Export}{Total amount of water diverted from the Delta by the Federal
and State governments to meet water agreements and contracts, expressed in
cubic feet per second (cfs). From Dayflow.}
\item{Temperature}{Water temperature in degrees Celsius from the
\code{discretewq} data package}
\item{Secchi}{Secchi depth in centimeters from the \code{discretewq} data
package}
\item{Salinity}{Salinity value on the Practical Salinity Scale from the
\code{discretewq} data package}
}
}
\source{
\itemize{
\item Sacramento Valley Water Year Index: \url{https://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST}
\item Dayflow: \url{https://data.cnra.ca.gov/dataset/dayflow}
\item DTO data from CDEC: \url{https://cdec.water.ca.gov/dynamicapp/staMeta?station_id=DTO}
\item \code{discretewq} data package (version 2.0.0): \url{https://github.com/sbashevkin/discretewq/releases/tag/v2.0.0}
}
}
\usage{
lt_seasonal
}
\description{
A long-term (1975-2021) integrated data set containing seasonal
averages of various metrics for each year and across the entire Delta. Used
in the analyses for the Drought Synthesis.
}
\keyword{datasets}
