% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_wq_1975_2021}
\alias{raw_wq_1975_2021}
\title{Raw values for water quality measurements for 1975-2021}
\format{
data frame with 34,523 rows and 14 columns
\describe{
\item{Source}{Name of the source dataset. Either EMP (DWR's Environmental
Monitoring Program), FMWT (CDFW's Fall Midwater Trawl survey), or STN
(CDFW's Summer Townet survey)}
\item{Station}{Location where measurement was collected}
\item{Latitude}{Latitude of \code{Station} in Decimal Degrees (WGS 84 Datum)}
\item{Longitude}{Longitude of \code{Station} in Decimal Degrees (WGS 84 Datum)}
\item{Region}{Region designation of \code{Station} as \href{https://github.com/mountaindboz/DroughtData/blob/master/data-raw/Rosies_regions.csv}{defined by the Drought Synthesis team}.
}
\item{SubRegion}{Subregion designation of \code{Station} as defined by the
\code{R_EDSM_Subregions_Mahardja_FLOAT} shapefile from the
\href{https://github.com/InteragencyEcologicalProgram/deltamapr}{\code{deltamapr}}
package.}
\item{YearAdj}{Adjusted calendar year - December-November, with December of
the previous calendar year included with the following year. For example,
December of 1974 is assigned year "1975" in this data set.}
\item{Season}{Winter (Dec-Feb), Spring (Mar-May), Summer (Jun-Aug), or Fall
(Sept-Nov)}
\item{Month}{Integer representing the month}
\item{Date}{Calendar date of the measurement}
\item{Datetime}{Date and time of the measurement. Not available for all
measurements.}
\item{Temperature}{Water temperature in degrees Celsius from the
\code{discretewq} data package. Data from EMP and FMWT collected in 2021 were
provided directly from data collectors and should be considered
provisional.}
\item{Salinity}{Salinity value on the Practical Salinity Scale from the
\code{discretewq} data package. Data from EMP and FMWT collected in 2021 were
provided directly from data collectors and should be considered
provisional.}
\item{Secchi}{Secchi depth in centimeters from the \code{discretewq} data
package. Data from EMP and FMWT collected in 2021 were provided directly
from data collectors and should be considered provisional.}
}
}
\source{
\code{discretewq} data package (version 2.1.0): \url{https://github.com/sbashevkin/discretewq/tree/1af04438c961c6d51f46cc777cbb6a142c1e3b30}
}
\usage{
raw_wq_1975_2021
}
\description{
A data set containing raw values of the water quality parameters
(water temperature, salinity, secchi depth) for 1975-2021. Seasonal and
Regional averages of this data are in the \code{lt_seasonal} and \code{lt_regional}
data sets, respectively. This data set contains only one data point per
\code{Station} and \code{Date} and excludes any data points that are more than 10
standard deviations away from the mean of each \code{SubRegion}. Used in the
analyses for the Drought Synthesis.
}
\keyword{datasets}
