% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_wq_1975_2021}
\alias{raw_wq_1975_2021}
\title{Raw values of water quality measurements for 1975-2021}
\format{
data frame with 34,523 rows and 14 columns
\describe{
\item{Source}{Name of the source dataset. Either EMP (DWR's Environmental
Monitoring Program), FMWT (CDFW's Fall Midwater Trawl survey), or STN
(CDFW's Summer Townet survey)}
\item{Station}{Location where measurement was collected}
\item{Latitude}{Latitude of \code{Station} in Decimal Degrees (WGS 84 Datum)}
\item{Longitude}{Longitude of \code{Station} in Decimal Degrees (WGS 84 Datum)}
\item{Region}{Region designation of \code{Station} within the larger Delta and
Suisun Marsh/Bay (Confluence, North, SouthCentral, Suisun Bay, Suisun
Marsh). The Drought Synthesis team categorized \code{SubRegion}s into broader
\code{Region}s using \href{https://github.com/mountaindboz/DroughtData/blob/master/data-raw/Rosies_regions.csv}{this crosswalk}
.}
\item{SubRegion}{Subregion designation of \code{Station} as defined by the
\code{R_EDSM_Subregions_Mahardja_FLOAT} shapefile from the
\href{https://github.com/InteragencyEcologicalProgram/deltamapr}{\code{deltamapr}}
package.}
\item{YearAdj}{Adjusted calendar year - December-November, with December of
the previous calendar year included with the following year. For example,
December of 2020 is assigned year "2021" in this data set.}
\item{Season}{Winter (Dec-Feb), Spring (Mar-May), Summer (Jun-Aug), or Fall
(Sept-Nov)}
\item{Month}{Integer representing the month}
\item{Date}{Calendar date of the measurement}
\item{Datetime}{Date and time of the measurement. Not available for all
measurements.}
\item{Temperature}{Water temperature in degrees Celsius}
\item{Salinity}{Salinity value on the Practical Salinity Scale}
\item{Secchi}{Secchi depth in centimeters}
}
}
\source{
\code{discretewq} data package (version 2.3.1):
\url{https://github.com/sbashevkin/discretewq/tree/v2.3.1}
}
\usage{
raw_wq_1975_2021
}
\description{
A data set containing raw values of three water quality
parameters (water temperature, salinity, and secchi depth) for 1975-2021.
Seasonal and Regional averages of this data are in the \code{lt_seasonal} and
\code{lt_regional} data sets, respectively. Used in the analyses for the Drought
Synthesis.

Data from 1975-2020 are from the \code{discretewq} data package. Data collected
in 2021 were provided directly from staff who collected the data and should
be considered provisional.

This data set contains only one data point per \code{Station} and \code{Date} and
excludes any data points that are more than 10 standard deviations away
from the mean of each \code{SubRegion}.
}
\keyword{datasets}
